+++ semantica/triplet_store/blazegraph_adapter.py
@@ -1,26 +1,27 @@
 """
-Virtuoso Adapter Module
+Blazegraph Adapter Module
 
-This module provides Virtuoso RDF store integration for RDF storage and
-SPARQL querying, supporting cluster connections and query optimization.
+This module provides Blazegraph integration for RDF storage and SPARQL
+querying, enabling connection to Blazegraph instances with namespace
+management and bulk loading capabilities.
 
 Key Features:
-    - Virtuoso connection and authentication
+    - Blazegraph connection and authentication
     - SPARQL query execution
     - Bulk data loading and management
-    - Graph and namespace management
-    - Cluster connection support
-    - Query optimization
+    - Namespace and graph management
+    - REST API integration
+    - Performance optimization
 
 Main Classes:
-    - VirtuosoAdapter: Main Virtuoso integration adapter
+    - BlazegraphAdapter: Main Blazegraph integration adapter
 
 Example Usage:
-    >>> from semantica.triple_store import VirtuosoAdapter
-    >>> adapter = VirtuosoAdapter(endpoint="http://localhost:8890/sparql", username="dba", password="dba")
+    >>> from semantica.triplet_store import BlazegraphAdapter
+    >>> adapter = BlazegraphAdapter(endpoint="http://localhost:9999/blazegraph", namespace="kb")
     >>> result = adapter.execute_sparql(sparql_query)
-    >>> load_result = adapter.bulk_load(triples, graph="http://example.org/graph")
-    >>> cluster_status = adapter.connect_cluster(cluster_config)
+    >>> load_result = adapter.bulk_load(triples)
+    >>> namespace_result = adapter.create_namespace("new_namespace")
 
 Author: Semantica Contributors
 License: MIT
@@ -34,106 +35,77 @@ import requests
 from ..semantic_extract.triple_extractor import Triple
 from ..utils.exceptions import ProcessingError, ValidationError
 from ..utils.logging import get_logger
+from ..utils.progress_tracker import get_progress_tracker
 
 
-class VirtuosoAdapter:
+class BlazegraphAdapter:
     """
-    Virtuoso RDF store adapter.
+    Blazegraph triplet store adapter.
 
-    • Virtuoso connection and authentication
+    • Blazegraph connection and authentication
     • SPARQL query execution
     • Bulk data loading and management
-    • Graph and namespace management
+    • Namespace and graph management
     • Performance optimization
     • Error handling and recovery
     """
 
     def __init__(self, endpoint: str, **config):
         """
-        Initialize Virtuoso adapter.
+        Initialize Blazegraph adapter.
 
         Args:
-            endpoint: Virtuoso endpoint URL
+            endpoint: Blazegraph endpoint URL
             **config: Additional configuration:
+                - namespace: Namespace name (default: "kb")
                 - username: Username for authentication
                 - password: Password for authentication
                 - timeout: Request timeout (default: 30)
-                - graph: Default graph URI
         """
-        self.logger = get_logger("virtuoso_adapter")
+        self.logger = get_logger("blazegraph_adapter")
         self.config = config
         self.progress_tracker = get_progress_tracker()
 
         self.endpoint = endpoint.rstrip("/")
-        self.username = config.get("username", "dba")
-        self.password = config.get("password", "dba")
+        self.namespace = config.get("namespace", "kb")
+        self.username = config.get("username")
+        self.password = config.get("password")
         self.timeout = config.get("timeout", 30)
-        self.default_graph = config.get("graph", "")
 
         self.connected = False
         self._connect()
 
     def _connect(self) -> None:
-        """Connect to Virtuoso instance."""
+        """Connect to Blazegraph instance."""
         try:
-            # Test connection with simple query
+            # Test connection
             sparql_endpoint = self._get_sparql_endpoint()
             response = requests.get(
                 sparql_endpoint,
                 params={"query": "SELECT * WHERE { ?s ?p ?o } LIMIT 1"},
                 timeout=self.timeout,
-                auth=(self.username, self.password),
+                auth=(self.username, self.password)
+                if self.username and self.password
+                else None,
             )
 
             if response.status_code == 200:
                 self.connected = True
-                self.logger.info(f"Connected to Virtuoso: {self.endpoint}")
+                self.logger.info(f"Connected to Blazegraph: {self.endpoint}")
             else:
                 self.logger.warning(
-                    f"Virtuoso connection test failed: {response.status_code}"
+                    f"Blazegraph connection test failed: {response.status_code}"
                 )
         except Exception as e:
-            self.logger.warning(f"Could not connect to Virtuoso: {e}")
+            self.logger.warning(f"Could not connect to Blazegraph: {e}")
 
     def _get_sparql_endpoint(self) -> str:
         """Get SPARQL endpoint URL."""
-        return urljoin(self.endpoint, "/sparql")
+        return urljoin(self.endpoint, f"/blazegraph/namespace/{self.namespace}/sparql")
 
-    def _get_sparql_update_endpoint(self) -> str:
+    def _get_update_endpoint(self) -> str:
         """Get SPARQL Update endpoint URL."""
-        return urljoin(self.endpoint, "/sparql-auth")
-
-    def connect_cluster(
-        self, cluster_config: Dict[str, Any], **options
-    ) -> Dict[str, Any]:
-        """
-        Connect to Virtuoso cluster.
-
-        Args:
-            cluster_config: Cluster configuration
-            **options: Additional options
-
-        Returns:
-            Connection status
-        """
-        # Virtuoso cluster connection
-        # This would typically involve multiple endpoints
-        endpoints = cluster_config.get("endpoints", [self.endpoint])
-
-        connections = []
-        for endpoint in endpoints:
-            try:
-                adapter = VirtuosoAdapter(endpoint, **cluster_config)
-                if adapter.connected:
-                    connections.append(endpoint)
-            except Exception as e:
-                self.logger.warning(f"Failed to connect to {endpoint}: {e}")
-
-        return {
-            "success": len(connections) > 0,
-            "connected_endpoints": connections,
-            "total_endpoints": len(endpoints),
-        }
+        return urljoin(self.endpoint, f"/blazegraph/namespace/{self.namespace}/sparql")
 
     def execute_sparql(self, query: str, **options) -> Dict[str, Any]:
         """
@@ -147,35 +119,36 @@ class VirtuosoAdapter:
             Query results
         """
         tracking_id = self.progress_tracker.start_tracking(
-            module="triple_store",
-            submodule="VirtuosoAdapter",
-            message="Executing SPARQL query on Virtuoso",
+            module="triplet_store",
+            submodule="BlazegraphAdapter",
+            message="Executing SPARQL query on Blazegraph",
         )
 
         try:
             if not self.connected:
                 self.progress_tracker.stop_tracking(
-                    tracking_id, status="failed", message="Not connected to Virtuoso"
+                    tracking_id, status="failed", message="Not connected to Blazegraph"
                 )
-                raise ProcessingError("Not connected to Virtuoso")
+                raise ProcessingError("Not connected to Blazegraph")
 
             sparql_endpoint = self._get_sparql_endpoint()
 
             self.progress_tracker.update_tracking(
-                tracking_id, message="Sending query to Virtuoso endpoint..."
+                tracking_id, message="Sending query to Blazegraph endpoint..."
             )
-            response = requests.get(
+            response = requests.post(
                 sparql_endpoint,
-                params={"query": query, "default-graph-uri": self.default_graph}
-                if self.default_graph
-                else {"query": query},
-                headers={"Accept": "application/sparql-results+json"},
+                data={"query": query},
+                headers={"Content-Type": "application/x-www-form-urlencoded"},
                 timeout=self.timeout,
-                auth=(self.username, self.password),
+                auth=(self.username, self.password)
+                if self.username and self.password
+                else None,
             )
 
             response.raise_for_status()
 
+            # Parse JSON response
             self.progress_tracker.update_tracking(
                 tracking_id, message="Parsing query results..."
             )
@@ -201,31 +174,6 @@ class VirtuosoAdapter:
             )
             raise ProcessingError(f"SPARQL query failed: {e}")
 
-    def optimize_queries(self, queries: List[str], **options) -> List[str]:
-        """
-        Optimize query performance.
-
-        Args:
-            queries: List of SPARQL queries
-            **options: Optimization options
-
-        Returns:
-            List of optimized queries
-        """
-        optimized = []
-
-        for query in queries:
-            # Basic optimization: remove unnecessary whitespace
-            opt_query = " ".join(query.split())
-
-            # Add LIMIT if not present (for SELECT queries)
-            if "SELECT" in query.upper() and "LIMIT" not in query.upper():
-                opt_query += " LIMIT 1000"
-
-            optimized.append(opt_query)
-
-        return optimized
-
     def bulk_load(self, triples: List[Triple], **options) -> Dict[str, Any]:
         """
         Load triples in bulk.
@@ -233,65 +181,52 @@ class VirtuosoAdapter:
         Args:
             triples: List of triples
             **options: Additional options:
+                - format: RDF format (turtle, ntriples, rdfxml)
                 - graph: Named graph URI
-                - format: RDF format (turtle, ntriples)
 
         Returns:
             Load status
         """
-        tracking_id = self.progress_tracker.start_tracking(
-            module="triple_store",
-            submodule="VirtuosoAdapter",
-            message=f"Bulk loading {len(triples)} triples to Virtuoso",
+        if not self.connected:
+            raise ProcessingError("Not connected to Blazegraph")
+
+        # Convert triples to RDF format
+        format = options.get("format", "turtle")
+        rdf_data = self._triples_to_rdf(triples, format)
+
+        # Upload endpoint
+        upload_endpoint = urljoin(
+            self.endpoint, f"/blazegraph/namespace/{self.namespace}/sparql"
         )
 
         try:
-            if not self.connected:
-                self.progress_tracker.stop_tracking(
-                    tracking_id, status="failed", message="Not connected to Virtuoso"
-                )
-                raise ProcessingError("Not connected to Virtuoso")
-
-            graph = options.get("graph", self.default_graph)
-            format = options.get("format", "turtle")
-
-            # Convert triples to RDF
-            self.progress_tracker.update_tracking(
-                tracking_id, message="Converting triples to RDF format..."
-            )
-            rdf_data = self._triples_to_rdf(triples, format)
-
             # Use SPARQL INSERT for bulk loading
-            update_endpoint = self._get_sparql_update_endpoint()
-
+            graph = options.get("graph", "")
             graph_clause = f"GRAPH <{graph}>" if graph else ""
+
+            # Build INSERT query
             insert_data = self._build_insert_data(triples)
             query = f"INSERT DATA {graph_clause} {{ {insert_data} }}"
 
-            self.progress_tracker.update_tracking(
-                tracking_id, message="Sending bulk load request..."
-            )
             response = requests.post(
-                update_endpoint,
-                data={"query": query},
+                upload_endpoint,
+                data={"update": query},
                 headers={"Content-Type": "application/x-www-form-urlencoded"},
-                timeout=self.timeout * 2,
-                auth=(self.username, self.password),
+                timeout=self.timeout * 2,  # Longer timeout for bulk operations
+                auth=(self.username, self.password)
+                if self.username and self.password
+                else None,
             )
 
             response.raise_for_status()
 
-            self.progress_tracker.stop_tracking(
-                tracking_id,
-                status="completed",
-                message=f"Bulk loaded {len(triples)} triples",
-            )
-            return {"success": True, "triples_loaded": len(triples), "graph": graph}
+            return {
+                "success": True,
+                "triples_loaded": len(triples),
+                "namespace": self.namespace,
+            }
         except Exception as e:
             self.logger.error(f"Bulk load failed: {e}")
-            self.progress_tracker.stop_tracking(
-                tracking_id, status="failed", message=str(e)
-            )
             raise ProcessingError(f"Bulk load failed: {e}")
 
     def _triples_to_rdf(self, triples: List[Triple], format: str = "turtle") -> str:
@@ -303,13 +238,9 @@ class VirtuosoAdapter:
                     f"<{triple.subject}> <{triple.predicate}> <{triple.object}> ."
                 )
             return "\n".join(lines)
-        else:  # ntriples
-            lines = []
-            for triple in triples:
-                lines.append(
-                    f"<{triple.subject}> <{triple.predicate}> <{triple.object}> ."
-                )
-            return "\n".join(lines)
+        else:
+            # For other formats, use simple turtle conversion
+            return self._triples_to_rdf(triples, "turtle")
 
     def _build_insert_data(self, triples: List[Triple]) -> str:
         """Build SPARQL INSERT DATA clause."""
@@ -356,7 +287,7 @@ class VirtuosoAdapter:
                     subject=binding.get("s", {}).get("value", ""),
                     predicate=binding.get("p", {}).get("value", ""),
                     object=binding.get("o", {}).get("value", ""),
-                    metadata={"source": "virtuoso"},
+                    metadata={"source": "blazegraph"},
                 )
             )
 
@@ -365,19 +296,21 @@ class VirtuosoAdapter:
     def delete_triple(self, triple: Triple, **options) -> Dict[str, Any]:
         """Delete triple."""
         if not self.connected:
-            raise ProcessingError("Not connected to Virtuoso")
+            raise ProcessingError("Not connected to Blazegraph")
 
-        update_endpoint = self._get_sparql_update_endpoint()
+        update_endpoint = self._get_update_endpoint()
 
         query = f"DELETE DATA {{ <{triple.subject}> <{triple.predicate}> <{triple.object}> }}"
 
         try:
             response = requests.post(
                 update_endpoint,
-                data={"query": query},
+                data={"update": query},
                 headers={"Content-Type": "application/x-www-form-urlencoded"},
                 timeout=self.timeout,
-                auth=(self.username, self.password),
+                auth=(self.username, self.password)
+                if self.username and self.password
+                else None,
             )
 
             response.raise_for_status()
@@ -386,3 +319,34 @@ class VirtuosoAdapter:
         except Exception as e:
             self.logger.error(f"Delete triple failed: {e}")
             raise ProcessingError(f"Delete triple failed: {e}")
+
+    def create_namespace(self, namespace: str, **options) -> Dict[str, Any]:
+        """
+        Create new namespace.
+
+        Args:
+            namespace: Namespace name
+            **options: Additional options
+
+        Returns:
+            Operation status
+        """
+        # Blazegraph namespace creation via REST API
+        create_endpoint = urljoin(self.endpoint, "/blazegraph/namespace")
+
+        try:
+            response = requests.post(
+                create_endpoint,
+                json={"namespace": namespace, **options},
+                timeout=self.timeout,
+                auth=(self.username, self.password)
+                if self.username and self.password
+                else None,
+            )
+
+            response.raise_for_status()
+
+            return {"success": True, "namespace": namespace}
+        except Exception as e:
+            self.logger.error(f"Create namespace failed: {e}")
+            raise ProcessingError(f"Create namespace failed: {e}")
