+++ semantica/utils/progress_tracker.py
@@ -121,7 +121,7 @@ class ConsoleProgressDisplay(ProgressDisplay):
             "semantic_extract": "🎯",
             "seed": "🌱",
             "split": "✂️",
-            "triple_store": "🗄️",
+            "triplet_store": "🗄️",
             "vector_store": "📊",
             "export": "💾",
             "reasoning": "🤔",
@@ -176,7 +176,7 @@ class ConsoleProgressDisplay(ProgressDisplay):
             "semantic_extract": "is extracting",
             "seed": "is seeding",
             "split": "is splitting",
-            "triple_store": "is storing",
+            "triplet_store": "is storing",
             "vector_store": "is indexing",
             "export": "is exporting",
             "reasoning": "is reasoning",
@@ -352,7 +352,7 @@ class JupyterProgressDisplay(ProgressDisplay):
             "semantic_extract": "🎯",
             "seed": "🌱",
             "split": "✂️",
-            "triple_store": "🗄️",
+            "triplet_store": "🗄️",
             "vector_store": "📊",
             "export": "💾",
             "reasoning": "🤔",
@@ -405,7 +405,7 @@ class JupyterProgressDisplay(ProgressDisplay):
             "semantic_extract": "is extracting",
             "seed": "is seeding",
             "split": "is splitting",
-            "triple_store": "is storing",
+            "triplet_store": "is storing",
             "vector_store": "is indexing",
             "export": "is exporting",
             "reasoning": "is reasoning",
@@ -901,7 +901,7 @@ class ProgressTracker:
             "semantic_extract": "🎯",
             "seed": "🌱",
             "split": "✂️",
-            "triple_store": "🗄️",
+            "triplet_store": "🗄️",
             "vector_store": "📊",
             "export": "💾",
             "reasoning": "🤔",
